requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
beanie==2.0.0
black==25.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dnspython==2.7.0
fastapi==0.116.1
flake8==7.3.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langsmith==0.4.8
lazy-model==0.3.0
mccabe==0.7.0
motor==3.7.1
mypy_extensions==1.1.0
numpy==2.3.2
openai==1.97.1
orjson==3.11.1
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pymongo==4.13.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyMuPDF==1.24.3
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
zstandard==0.23.0
